[SH] Remove support for some (older) STMicroelectronics Boards
This update simply removes support from some STMicroelectronics Boards
that are older, and no longer required to be supported in the current
STLinux wavefront for U-Boot.  The boards in question are (mostly) based
on STMicroelectronics SoCs using either the SH4-200 or ST200 series cores.

The following boards have been removed in this commit:

	SoC:		Boards
	----		------

	STb7100/9	MB411, MB442, MB448, HMP7100, HMS1

	STx7200		MB519, CB101, CB102, CB103

	STi5301:	MB390, MB424

	STm8010:	STm8010-MBoard, TRAVIATA

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>