[SH] Remove support for some (older) STMicroelectronics Boards (part 2)
This update is a follow-on to the previous commit, which removed support
from some STMicroelectronics Boards that are older, and no longer
required to be supported in the current STLinux wavefront for U-Boot.
The boards in question are (mostly) based on STMicroelectronics SoCs
using either the SH4-200 or ST200 series cores.

The previous commit largely just "deleted" (what are now) unwanted files,
this update edits the kept files to remove references to "old" boards etc.

All references to deleted boards/SoCs have been removed in this commit:

	SoC:		Boards
	----		------

	STb7100/9	MB411, MB442, MB448, HMP7100, HMS1

	STx7200		MB519, CB101, CB102, CB103

	STi5301:	MB390, MB424

	STm8010:	STm8010-MBoard, TRAVIATA

NOTE:	This commit has removed *all* references to the CPU Architecture
	for STMicroelectronics' ST200 series of cores, and the ST231.

NOTE:	Only a few vestigial references to the SH4-100 and SH4-200 series
	of CPU cores remain, as they may still prove useful!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>