[ST40] Globally renamed the architecture+CPU from "sh" to be "st40"
This update logically does not change the code, it just renames the name
of the "architecture" and the "CPU" from "sh" to be "st40".  In addition,
the following files/directories have consequently been renamed:

	renamed:    cpu/sh/         ->  cpu/st40/
	renamed:    include/asm-sh/ ->  include/asm-st40/
	renamed:    lib_sh/         ->  lib_st40/
	renamed:    sh_config.mk    ->  st40_config.mk

The extant baseline for STMicroelectronics' version of U-Boot is the
"v1.3.1" release. Historically, STMicroelectronics added support for
various ST40 (SuperH SH-4) based SoCs to this "v1.3.1" release. For
these STMicroelectronics' systems, both the "architecture" and the
"CPU" names were specified as "sh", as this was consistent with the
linux kernel source trees.

However, in the official "v1.3.2" U-Boot release, other
(non-STMicroelectronics) developers added support for other
(non-STMicroelectronics) "sh" architecture+CPUs. As a result of
this parallel development, with different trees with the same "sh"
identifiers, the two trees were maintained separately, and as a result
the STMicroelectronics' tree has not been subsequently re-based.

There is now a requirement to have a single unified tree, which
supports both the non-STMicroelectronics "sh" products, as well as the
STMicroelectronics "sh" products. However, in order to re-baseline the
STMicroelectronics' tree, it has been decided to rename (temporarily)
the architecture+CPU names from "sh" to "st40" in the STMicroelectronics'
tree.  Then, once the re-baseline is complete, it should be easier to
merge the final "sh" and "st40" trees into a unified "sh" tree. (Well,
that is the theory!)

This update, therefore, is just to move the "sh" domain into the "st40"
domain, in the STMicroelectronics' tree, so that the aforementioned
re-baselining can proceed accordingly.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>