[ST40] Globally renamed the CONFIG_xxx names from "SH" to be "ST40"
This update logically does not change the functionality of code, it just
renames the names of the CONFIG_xxx and CFG_xxx configuration macros from
the "SH" name-space to the "ST40" name-space. Please see the previous
commit for rationale and an explanation.

In essence, the following configuration macros have been changed:

		Old Name		New Name
		--------		--------
		CFG_SH_xxx	->	CFG_ST40_xxx
		CONFIG_SH_xxx	->	CONFIG_ST40_xxx
		CONFIG_SH4	->	CONFIG_ST40

This update, therefore, is just to move the configuration macros from
the "SH" domain into the (previously unused) "ST40" domain, in the
STMicroelectronics' tree.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>