[ST40] Remove support for the "SCIF" UART for ST40 based systems.
This update removes support for the legacy "SCIF" UART, which on
STMicroelectronics' ST40-based systems now uses the "ASC" UART instead.
Note, support for the SCIF still exists in the (Renesas SuperH) "SH" tree,
support has only been removed from the (STMicroelectronics) "ST40" tree.

This update should be considered as logically part of an earlier commit,
commit c24611ae67c3dfa0eeb7f5dd4ec2e28012ee5f95 which removed support
for some (older) STMicroelectronics Boards. It was at that juncture that
ST40-based systems no longer needed support for the SCIF.

In addition, all remaining references to the macro CONFIG_SH_SCIF_SERIAL
have been removed, as this was specific to the SCIF in a ST40 context,
which is no longer supported in the STMicroelectronics' environment.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>