[SH] Fix RMII "re-timing" configurations for GMAC on the STx7108
This update fixes one of the "re-timing" configurations, specifically
for RMII mode, as used on the ST-GMAC on STMicroelectronics' STx7108 SoC.

The previous versions of the "re-timing" configurations were (in some
cases) incorrect, and may have only worked by fluke! When attempting
to add support for the B2037A board, it was noticed that the RMII
configurations needed to be re-implemented to get RMII to work correctly.
The relevant "Application Note" for the re-timing logic was recently
modified (mainly for RMII), and the previous code was based on an
older version of this Application Note.  This update includes the
re-implementation for the RMII bus re-timings (only), on the STx7108,
using the latest version of the relevant Application Note.

Note:	It is believed that all the other MII bus configurations will
	need to undergo a similar "re-implementation" activity (or at
	least an analysis to confirm their extant correctness).

Thanks to Srinivas Kandagatla, who provided the fixed re-implementation
for linux, which is the basis of this commit (transliterated by Sean
McGoogan).

All testing was on a single board, "STIH205-HDK10ST N011", together with
the code in the *next* commit, which provides a working RMII ethernet
support for the B2037A, using the IC+ IP101A PHY.

Signed-off-by: Srinivas KANDAGATLA <srinivas.kandagatla@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>