[ENV] fixed ability to generate the target tool "fw_printenv"
Some time ago, the ability to build the target environment tool
"fw_printenv" (by default, put in tools/env/fw_printenv) was broken.
This commit fixes that issue, and now allows the target-side environment
tools to be successfully built (cross-compiled) on the host system.

Testing was restricted to building the target binary for one of the
STMicroelectronics board, i.e.:

	host% make distclean
	host% make hdk7108se_config
	host% make env

See also: https://bugzilla.stlinux.com/show_bug.cgi?id=12718

Thanks to Jon for the patch (in the above bugzilla), which was was
applied with only a few changes by the committer.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>