[ENV] fix annoying compiler warnings in "fw_env.c"
This commit fixes the latent compiler warnings that were present in
tools/env/fw_env.c, now that the previous commit allows us to compile
it properly!

Essentially, it was just adding some "(char*)" and "(uchar*)" casts in
the appropriate place to ensure the sanctity of "signed-correctness"
is preserved ...

No testing was done on this commit, as logically no code was changed!
Testing was limited to ensuring that it now builds cleanly, at least on:
	sh4-linux-gcc (GCC) 4.2.4 (snapshot) (STMicroelectronics/Linux Base 4.2.4-71)
	sh4-linux-gcc (GCC) 4.5.2 20110124 (STMicroelectronics/Linux Base 4.5.2-78)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>