Remove inferior (historical) bug-fix to fs/cramfs/cramfs.c
This commit undoes one bug-fix, for which a separate (and better)
bug-fix was independently provided.

Commit b8e8d599e5a0879c2ed48228b1c1c3354e730194 fixed an issue in the
file fs/cramfs/cramfs.c which prevented this file compiling, if the
macro CFG_NO_FLASH was defined.

However, commit f540c42d9564854b19ce9bbb70affe172529fe70 was a better fix.

During the recent merge (ST40 + v1.3.2) both of the fixes were enacted,
as they could both co-exist. This update is simple to remove the inferior
fix, leaving only the superior fix in place.

As a result, the file fs/cramfs/cramfs.c is now identical to the one
tagged as part of the v1.3.2 release.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>