[CFI] Fixes to cfi_flash.c from STMicroelectronics' ST40 tree
As mentioned in commit 56c8196349ed7b8cc54baabea31b359e462216d2,

>	2) drivers/mtd/cfi_flash.c
>	This file has been extensively modified between 'v1.3.1' and
>	'v1.3.2'. In addition, there have been numerous changes to this
>	file when it was in the STMicroelectronics' ST40 tree. As a
>	result, I decided to use the unmodified v1.3.2 version (ignoring
>	all the changes in the STMicroelectronics' ST40 tree). Hence,
>	post this commit, the "lost" commits will need to reviewed, and
>	(optionally) re-applied. This will be done as a separate commit.

In fact, there are a total of 8 commits that were applied to the
"drivers/mtd/cfi_flash.c" file, in the STMicroelectronics' ST40 tree.
This commit includes the coalescing of the first four of them, the
remaining four will be added shortly after this commit, at which juncture
the merge of the 'v1.3.2' into the STMicroelectronics' ST40 tree will
have been deemed to have been completed. The first 4 are included as a
single "coalesced" commit, as they are all bug-fixes (mostly retaining to
'reset') and they tend to interact with each other.

1) Fix to CFI flash, to support multi-bank devices better.
2) [CFI] Further reliability improvements to NOR Flash CFI handling.
3) [FIX] Bug fix for CFI FLASH devices.
4) [Fix] Fix incorrect probing of AMD (compatible) CFI flash devices.

1) see commit 7e87c290d9617d3743846495247331bae8daabe9
2) see commit 1bfd9de379f2a9dfd283a2ed295819843253367e
3) see commit 0878d5479f679ffe738f9f410af8421d67872990
4) see commit 19f85e74e70a7f512c527c08d6aafce7266695fc

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>