[ST40] Complete Merge of 'v1.3.2' into STMicroelectronics' ST40 tree
This commit completes the merge of the official 'v1.3.2' release (which
includes Renesas SuperH sh3 and sh4 trees) and the STMicroelectronics'
ST40 tree. The two trees are in different name spaces, and hence should
not occlude or conflict with each other.

Specifically, *all* the delinquencies are now (believed to be) addressed,
and there are no (known) outstanding issues with the merge to 'v1.3.2'.

see commit 56c8196349ed7b8cc54baabea31b359e462216d2

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>