[ST40] Remove support for SMC911x driver (from original ST40 tree)
This update is a pre-cursor to the merge with v1.3.3 on the ST40 tree.

Without this commit, there would be a "add/add" GIT conflict to content
with, during the next merge.  That conflict is due to the fact the file
"drivers/net/smc911x.c" was added to the STMicroelectronics' tree, and was
also independently added in the v1.3.3 release. However, when some (older)
STMicroelectronics' boards were removed, the STMicroelectronics version of
the driver should have been removed at the same time (but was not), as the
driver was not used by any of the *remaining* STMicroelectronics' boards.
Hence, removing this driver does not change the functionality of U-Boot.

In summary, this commit removes an unused driver, by deleting the
drivers/net/smc911x.{c,h} files, which (in hindsight) should have probably
been done in commit c24611ae67c3dfa0eeb7f5dd4ec2e28012ee5f95, and also
modifies drivers/net/Makefile, which (in hindsight) should probably have
been done in commit 65491ebe094ae930c966103b2f3dcff270496ec7.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>