[SH,FIX] Preserve MMUCR.SE when setting MMUCR.TI on SH4-200 cores
With reference to commit 38533e4dc772e3dd30ceeb92480a79957355cff7,
where it was required to invalidate the ITLB, otherwise problems will
occur if GDB had setup a 512 MiB page in PMB slot #0.

It has now transpired that the aforementioned commit accidentally
broke boot-from-GDB on ST40 SoCs, which used the older SH4-200 series
cores. There is no problem if booting from FLASH, or using any of the
newer SH4-300 series cores.

This code solves this issue by preserving the value of the MMUCR.SE bit,
when we set the MMUCR.TI bit, to invalidate the ITLB after we re-program
the PMB for SH4-200 series cores. In the case of SH4-300 series cores,
we actually set the bit-field to 0, but this is benign, as the SE bit
is located in a different register (and hence not accidently corrupted
when we write to the MMUCR.TI bit-field).

This was tested on a single MB442 board, with a STx7109c3 (SH4-200) SoC,
both boot-from-NOR and boot-from-GDB, when built in 32-bit (SE) mode.

Note: When using a system where GDB will *not* configure a 512MiB slot in
the PMB, then an alternative solution is possible - comment out the poke
to the MMUCR, i.e. un-do commit 38533e4dc772e3dd30ceeb92480a79957355cff7.
Given that there are probably no boards with 512MiB + SH4-200 available,
then another simple option would be to only perform the ITLB invalidation
on SH4-300 parts, i.e.:
	#if defined(CONFIG_CPU_SUBTYPE_SH4_3XX) /* it is an SH4-300 */
	invalidate_itlb:
		MOV_CONST32_R0 SH4_CCN_MMUCR
		mov	#(SH4_MMUCR_TI), r1
		mov.l	r1, @r0			/* MMUCR.TI = 1 */
	#endif
However, the submitted solution was chosen as a "belt and braces"
solution which should work in *all* possible possible situations, and
does not make any (potentially) unsafe assumptions.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>