[ST40] Reintroduce support for "bootm" on ST40 SoCs
As mentioned in commit 85ccb41620140f19ad06176c6b4c7395659845f9:

>	There is one main functional delinquency from this commit:
>
>	1) common/cmd_bootm.c and lib_st40/bootm.c
>
>	Support for the "bootm" command had *not* been updated, to take
>	into account the new "image" infrastructure. This clearly needs
>	to be added, but probably best done as a separate commit.

This commit addresses this delinquency, and now reintroduces support for
the "bootm" command (at least for "legacy" images) for the ST40 SoCs. It
should be noted that the new-format images are not yet supported.

This has been tested by burning encapsulated vmlinux + ramdisk images into
(NOR) FLASH, and booting linux successfully. Only "legacy" images were
tested, with the kernel being tried in both compressed and uncompressed
forms.  All testing was performed on a single STi7108-HDK10ZA-N206 board.

In addition, the boot_get_ramdisk() function has been used to load
the initrd RAM disk. As a result, it should be noted that the following
environment variables may now be used to control placement of the ramdisk
in RAM, as passed to linux:

		bootm_low, bootm_size, initrd_high

See the "README" for further details of these environment variables.

Finally, the influence of the macro CFG_BOOTMAPSZ has now logically been
removed from the "bootm" command (for ST40). This macro often caused
users a lot of trouble, especially if using a large RAM disk, or if
.zero_empty_page was moved higher than "normal". As well as removing
sensitivity on the value of CFG_BOOTMAPSZ, the initrd should now (by
default) be relocated to ~64KiB below the end of the stack. Also, is
it now possible to force U-Boot *not* to relocate the initrd at all,
by doing the following:
	U-Boot> setenv initrd_high ffffffff

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>