[ST40] Remove CFG_BOOTMAPSZ from all ST40 board configuration files.
As mentioned in the previous commit, the influence of the macro
CFG_BOOTMAPSZ has now logically been removed from the "bootm" command
processing (for ST40). This commit strengthens that statement by
physically removing the macro's definition in the *all* the ST40 <board>.h
configuration files. Thus, the macro is not even seen by the compiler!

The behaviour of this macro can now be achieved much more more flexibly,
using the following environment variables instead:

		bootm_low, bootm_size, initrd_high

See the "README" for further details of these environment variables.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>