[SH] Remove CFG_BOOTMAPSZ from all ST40 board configuration files.
The influence of the macro CFG_BOOTMAPSZ has now been removed from the
"bootm" command (for ST40). This macro often caused users a lot of
trouble, especially if using a large RAM disk, or if .zero_empty_page
was moved higher than "normal".

As well as removing sensitivity on the value of CFG_BOOTMAPSZ, the initrd
should now always be relocated to ~64KiB below the end of the stack.
Which, for all ST40 boards should be a valid and safe change, as long as:

	CFG_SDRAM_BASE << stack < TEXT_BASE < CFG_SDRAM_BASE+CFG_SDRAM_SIZE
	and CFG_SDRAM_SIZE is "suitably large".

This has been tested by burning encapsulated vmlinux + ramdisk images into
(NOR) FLASH, and booting linux successfully, with the kernel being tried
in both compressed and uncompressed forms.  All testing was performed
on a single STi7108-HDK10ZA-N206 board.

Also see: https://bugzilla.stlinux.com/show_bug.cgi?id=13298

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>