[SH] Add support for "initrd_high" environment variable for ST40 systems
The previous commit removed CFG_BOOTMAPSZ from all ST40 board
configuration files. However, this change resulted in one special case of
booting linux not being supported properly. That case was when the linux
sees *less* memory than U-Boot, and is often encountered when users pass
"mem=xxM" as a command line option to the linux kernel.

A simple workaround for this issue is just to change CFG_SDRAM_SIZE to
agree with the amount of RAM that linux sees, and re-build and re-flash
U-Boot. U-Boot will then load the initrd just under that stack (as
normal), and everything should now be consistent.

A better means to address this issue, is to honour the semantics of
the environment variable "initrd_high" for ST40 systems. This has the
advantage of being "dynamic" inasmuch as it is not needed to re-compile,
nor re-flash U-Boot to change this limit.

Please see "README" for details of the semantics of this. In summary:

	1) not defined, or set to "no", or evaluated to be 0x0:

		Relocate initrd just under the stack (as before).
		i.e. feature disabled.

	2) set to 0xFFFFFFFF:

		Do no relocate the initrd at all. i.e. use "in situ".
		It is the user's responsibility to ensure the initrd is
		located at an appropriate address for linux & U-boot!

	3) set to some other address:

		Relocate, but limit the upper-bound of the relocated
		initrd to the given address.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>