[SH] Improve support for "initrd_high" environment variable on ST40 systems
The previous commit added support for the "initrd_high" environment
variable for ST40 systems. However after further empirical testing it
appears that "very last" page that linux uses is "special" in the boot
process, and should not be used to hold an "initrd".

So, for example, on a 32-bit system with 128MiB of SDRAM, with "mem=64M"
(say) passed to linux, then one would reasonably expect to use:

	U-Boot> setenv initrd_high 84000000

However, this would physically place the initrd at the end of memory,
including the page 0x83fff000..0x83ffffff - this causes problems for
linux!  A simple workaround is just to move the initrd_high down by one
page, i.e.

	U-Boot> setenv initrd_high 83fff000

This will ensure that the last page (as seen by linux) is not used by
U-Boot to relate the initrd, and solves the aforementioned problem.

However, this "off-by-one" page requirement is not intuitive, and will
ultimately lead to users being confused. In an effort to make it more
"intuitive", this commit is a simple workaround which simply will not use
the last page, by using the penultimate page instead, as the upper-bound
when relocating the initrd.

Hence, it is now possible (in this example) to use both "mem=64M" and
initrd_high=84000000 and U-Boot will not relocate any part of the initrd
to the last page as seen by linux (0x83fff000..0x83ffffff).

This update does not in any way modify the semantics when "initrd_high"
is set of 0xFFFFFFFF. It should be noted that users can still use this
"trick" to force a ramdisk on the last page - caution is advised against
doing this (either accidentally or deliberately)!

This has been tested by burning encapsulated vmlinux + ramdisk images into
(NOR) FLASH, and booting linux successfully, with the kernel being tried
in both compressed and uncompressed forms.  All testing was performed on
a single STi7108-HDK10ZA-N206 board. Testing was performed with passing
"mem=64M", with various values of "initrd_high" used. No boot problems
were noted.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>