Add "stdint.h" which appears to be missing on some cross toolchains
With reference to commit 999b4ac82bf04e01d81c1299ddd8621fd09ac072,
and to commit 9f46b8b093524849d409a04222b68656154a3c8d, as a result of
performing a fuller regression testing, a new problem was noticed that
only affected *some* boards, and only with some toolchains!

Specifically, the problem was only noticed on ST40 boards that enabled
support for NAND, and only when using the LDDE 2.3 toolchain.

This commit fixes this problem essentially by including a local copy of
"stdint.h" in U-Boot's "include" directory. As a result, all ST40 boards
now build cleanly, with both of the following toolchains (2.3 + 2.4):

	sh4-linux-gcc (GCC) 4.2.4 (snapshot) (STMicroelectronics/Linux Base 4.2.4-71)
	sh4-linux-gcc (GCC) 4.5.2 20110124 (STMicroelectronics/Linux Base 4.5.2-78)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>