[SH] Re-factor code for NAND on STMicroelectronics' SoCs
This commit is an initial attempt at re-factoring the code to support
NAND on STMicroelectronics' SoCs to improve both maintainability and
extensibility - mainly extensibility of varying ECC schemes.

Specifically, the code pertaining to STMicroelectronics' "FLEX-mode" has
been moved into cpu/sh/stm-nand-flex.c. Also, some of the functions and
structures have had their scope reduced from "extern" to "static". In
addition a new function stm_nand_init() has been created as a single
place in which to initialize the "hybrid" (pair of) ECC schemes.

In addition, the following 4 combinations are now all supported
(previously only 2 of them were):

	Driver		ECC Schemes
	------		-----------
	Bit-Banging	S/W 3/256
	Bit-Banging	S/W 3/256 + H/W 3+1/128 "Boot-Mode"
	FLEX-mode	S/W 3/256
	FLEX-mode	S/W 3/256 + H/W 3+1/128 "Boot-Mode"

Note: it is not obvious there is any need for Bit-Banging + "Boot-Mode"!

The rationale for the above, is to ease the immanent addition of support
for the STMicroelectronics' "AFM4" (Advanced Flex Mode) 4+3/512 ECC
scheme in U-Boot.

Note: ECC schemes of the form "x+y/z", means x bytes of ECC + y bytes of
"tag/ID" per z byte record of data.

All testing was performed on a single HDK-7108 (N206) board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>