[SH] Re-factor code for NAND on STMicroelectronics' SoCs (part 2)
This commit builds on the previous commit, which was an initial attempt
at re-factoring the code to support NAND on STMicroelectronics' SoCs to
improve both maintainability and extensibility - mainly extensibility
of varying ECC schemes.

This update now includes a new ST-specific NAND initialization function
stm_default_board_nand_init() which the board-specific function
board_nand_init() can simply call. This means that the board-specific
function can simply just call the "default" initialization function,
and only that one default function needs to be maintained, as new NAND
capabilities are added. As a result, all the "nand.c" files in the ST
boards directory have been modified to call stm_default_board_nand_init()
to perform the bulk of the NAND initialization in one place.

Also, the function nand_init_chip() in drivers/mtd/nand/nand.c has
been updated to pass the (mtd_info) pointer to 'mtd' in the 'priv'
field of the nand_chip structure, to board_nand_init(), and hence
to stm_default_board_nand_init(). The benefit of this is that the
stm_default_board_nand_init() now has access to both the 'mtd'
and 'nand' pointers, which allows it to initialize both structures
*properly*. Hence, the previous "bodge" of hijacking the 'select_chip'
function pointer to initialize the 'mtd' structures on first invocation
(then perform the normal function) can be safely removed. As a result, the
initialization code is now cleaner and simpler, and the hijacking bodge
has now be removed.  This should all improve clarity, maintainability,
and extensibility.

All testing was performed on a single HDK-7108 (N206) board, as before.
In addition, ./MAKEALL was run for all the ST boards, to ensure that
everything builds cleanly.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>