[SH] Preparations for adding STM's propitiatory "AFM4" NAND ECC scheme
This update is in preparation for adding support for STMicroelectronics'
"AFM4" ECC scheme, as used with SLC NAND devices on various STM SOCs.

This scheme uses 26 (24+2) bits (4-bytes) to store the ECC, to protect
each 512-byte record. In addition, a 3-byte "marker" is stored alongside
the 4-byte ECC in the OOB, hence this scheme takes a total of 7 bytes
to protect each 512 byte record.

Essentially, this update is just making some of the present ECC
infrastructure ready to work with the (to be added) 4+3/512 ECC scheme.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>