net/designware: cleanup code to check the result of auto-negotiation
Code was bit complex & error prone to check the conditions. There is
one bit to check speed & duplex in PHY registers e.g
   - bit 6 of PHY register 5 SET means device is capable of 10/full
     duplex
   - bit 5 of same register means 10/half duplex.

This patch modified the code in such a way that we check this bit &
configure variables accordingly. Earlier we were checking speed bits to
know the speed & then duplex.

Signed-off-by: Vikas Manocha <vikas.manocha@st.com>