[SH] Add support for STM's propitiatory "AFM4" NAND ECC scheme
This update is to add support for STMicroelectronics' "AFM4" ECC scheme,
as used with SLC NAND devices on various STM SoCs.

This scheme technically uses 26 (24+2) bits (4-bytes) to store the ECC,
to protect each 512-byte record. In addition, a 3-byte "marker" is stored
alongside the 4-byte ECC in the OOB, hence this scheme takes a total of
7 bytes to protect each 512 byte record.

The STMicroelectronics' "AFM4" (Advanced Flex Mode) scheme is referred to
as being a "4+3/512 ECC" scheme.  Where: ECC schemes of the form "x+y/z",
means x-bytes of ECC plus y-bytes of "marker" per z-byte record of data.

It should be noted that although U-Boot now understand the AFM4 ECC
scheme, U-Boot does *not* use the real H/W AFM mode. Instead, you
still need to use either the S/W "bit-banging", or the H/W "FLEX-mode"
drivers. Hence, the AFM4 ECC is supported purely as a S/W emulation
of the H/W AFM ECC engine.  It is simply the addition of the AFM4
ECC algorithm+layout that is being added here.  As a result, it is
now possible to use H/W AFM on linux (with AFM4 ECC), and it will be
compatible with U-Boot's usage.

It is also possible to use the AFM4 ECC *with* the 3/128 "boot-mode"
ECC, on the same SLC NAND device. The two can happily co-exist. Please
set the following macro to define the boundary where the boot-mode ECC
stops, and the AFM4 ECC starts:

	#define CFG_NAND_STM_BOOT_MODE_BOUNDARY ...

In addition, the following 8 combinations are now all supported in U-boot:

	Driver		ECC Schemes
	------		-----------
	Bit-Banging	S/W 3/256
	Bit-Banging	S/W 3/256	+ H/W 3+1/128 "Boot-Mode"
	Bit-Banging	AFM4 4+3/512
	Bit-Banging	AFM4 4+3/512	+ H/W 3+1/128 "Boot-Mode"
	FLEX-mode	S/W 3/256
	FLEX-mode	S/W 3/256	+ H/W 3+1/128 "Boot-Mode"
	FLEX-mode	AFM4 4+3/512
	FLEX-mode	AFM4 4+3/512	+ H/W 3+1/128 "Boot-Mode"

Note: it is not obvious there is any need for Bit-Banging + "Boot-Mode"!

All testing (for all 8 combinations above) was performed on a single
HDK-7108 (N206) board, as before.  In addition, ./MAKEALL was run for
all the ST boards, to ensure that everything builds cleanly.

It should be noted that by default, U-Boot has disabled the newly added
AFM4 ECC scheme, and so be 100% compatible "out of the box" with previous
releases. To enable this new capability, then the following macro should
be defined in your board-specific config.h file.

	#define CFG_NAND_ECC_AFM4	/* for AFM4 (4+3/512) ECC */

Finally, it should be noted that the AFM4 ECC is incompatible with the S/W
3/156 ECC the was previously used. Scrubbing the NAND flash device clean
when migrating between the two incompatible schemes is highly recommended.
In particular, please note that when using a NAND-resident Bad Block Table
(BBT), that the BBT is *also* protected by the active ECC scheme. Hence,
when migrating between these ECC schemes, you may get warnings about :-
"ECC error while reading bad block table". This is indicative that the
BBT was created with a different ECC scheme. Again, scrubbing the NAND
device should rectify this situation.

A word of caution - if you are using the "nand scrub" command, please
take a note of the original list of bad blocks, and then mark them as
bad (modulo the BBTs itself) afterwords (nand markbad), otherwise,
some genuinely bad blocks may (erroneously) now appear to be "good"
(which would be bad)!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>