[SH] Refactored code for (better) GPIO handling STx7108 (et seq.)
This commit is a pre-cursor for adding support for STMicroelectronics'
B2000 board which uses the new STxH415 (Orly) SoC. Since the STx7108,
all new SoCs appear have a more uniform approach to managing the GPIOs
and their alternative function capabilities in the H/W. As a result of
preparing for the STxH415 B2000 platform, a better (more consistent) means
of configuring the PIOs in S/W was desired, and this was retrospectively
applied to the STx7108 (viz, this commit).

The current approach included in this commit should be simpler, cleaner,
more consistent, and should be used for all new SoCs (STx7108 and later).
This commit is specifically only for the STx7108 SoC, and the relevant 4
supported boards (HDK7108, MB837, MB903, and B2037). As well as providing
the common structures for subsequent users of this solution.

The main change is the removal of the stx7108_pioalt_pad_cfg structure,
and replacing it with the enumeration type stm_pad_gpio_direction.
The old stx7108_pioalt_pad_cfg paradigm allowed each of the 3 enables
(OE, PU, OD) to be controlled individually, which was providing too
much flexibility and complexity, when there are only 8 theoretical
possibilities, only 5 of which the H/W sensibly supports. So, now
each PIO pin is associated with exactly one enumeration value, and the
bit-twiddling functions know exactly what to do with the low-level 3
enable signals, for that one enumeration value.

It should also be noted that the previous scheme used 2-bits (signed)
for each of the three enables (OE, PU, OD). Only 3 values were supported:
		negative == do not change it
		zero     == disable it
		one      == enable it
However, is hindsight, the "do not change" option on an individual
enable had questionable benefit, and has now been removed. Hence, all
3 enables will be explicitly programmed, unless the enumeration value
"stm_pad_direction_ignored" is used, in which case no programming will be
performed at all. The value "stm_pad_direction_unknown", should never be
used, other than as an initial value, which will be over-written prior
to the proper PIO configuration taking place.

Finally, it should be noted that aforementioned enumeration
type paradigm is *only* for PIOs using alternative functions not
equal to zero (i.e. alternative functionality, and not just the
simple/traditional GPIOs). Please see the enumeration definition
in "include/asm-sh/pio-control.h", for further details of the PIO
configurations.

It should be noted, that apart from compiling it (cleanly), it has *not*
been tested on any STx7108 boards. Regression testing will be done once
support for the STxH415 is added.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>