[SH] Explicitly enable the Hamming controller when needed
It is required that either the Hamming or the BCH controller is enabled,
as appropriate, before attempting to use it. The two register spaces of
the two controllers occlude each other. Talking to the wrong one will
result in undefined (typically very bad) behaviour!

There is a configuration bit which controls which controller is enabled.
It is essential to set this bit appropriately to enable the correct
controller *before* performing *any* FLEX-mode register accesses.

It should be noted that, the STxH415 when booting from SPI serial flash,
by default, actually configures this bit to enable the BCH (and not the
Hamming) controller! As a result, attempting to use FLEX-mode accesses to
talk to the Hamming controller, when the BCH is enabled must be avoided!

To force U-Boot to enable the Hamming controller (and hence disable the
BCH controller), then please define the following macro:

	#define CFG_SH_NAND_USE_HAMMING

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>