[SH] Add some 'helper' PIO macros (to allow pairing of PIO pin values)
This commit just adds a few helper macros which allow PIO pins to be
defined, referenced and programmed much easier. PIO pins are defined by
a pair of values: the PORT number, and the BIT number (0-7). Often these
two numbers should be treated logically as a single tuple (or 'pair'),
which uniquely defines that specific PIO pin.

Ideally, it would be useful if a single #define macro could be used
to refer to this unique pair of values, as a single logical parameter.
This commit adds 3 new macros which allow such 'pair' definitions to be
used relatively easily. The three new macros are:

		#define SET_PIO_PIN2(PAIR, DIR)
		#define STPIO_SET_PIN2(PAIR, V)
		#define STPIO_GET_PIN2(PAIR)

This now allows the following simpler code sample to be written:

{
	#define SPI_MISO	12, 4		/* PIO12[4] */
	#define SPI_MOSI	12, 5		/* PIO12[5] */

	/* set PIO directionality */
	SET_PIO_PIN2(SPI_MISO, STPIO_IN);	/* SPI_MISO */
	SET_PIO_PIN2(SPI_MOSI, STPIO_OUT);	/* SPI_MOSI */

	/* drive outputs with sensible initial values */
	STPIO_SET_PIN2(SPI_MOSI, 0);		/* deassert SPI_MOSI */
}

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>