[SH] Added initial support for B2000 (STxH415) Validation Board
This commit adds initial support for STMicroelectronics' B2000 Validation
Board, which uses STMicroelectronics' STxH415 (Orly) SoC device.

The following features have now been successfully tested:

	* 32-bit (SE) mode on the ST40 CPU
	* Serial ASCs (3 off)
	* SPI Serial Flash (FSM and "bit-banging")
	* SLC NAND Flash (FLEX-mode)
	* Boot from SPI Serial Flash
	* Boot from NAND Flash
	* Ethernet (MII, using B2032A daughter-board)
	* I2C bus

The following features are to be completed:

	* ARM CPU support
	* RMII Ethernet (using B2035A daughter-board)
	* USB (OHCI)
	* SATA

The following features are not planned to be supported:

	* 29-bit mode
	* NOR flash (no NOR on the board)

Note, using an HDMI TV (CN5, HDMITX) to test I2C, the I2C bus was running
at 100kHz (i.e. "standard mode, and not 400kHz) as required by the HDMI
specification.

All testing was on a single board, "B2000B-00 N026", expect for the I2C
testing, which was performed on "B2000C N011".

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>