[STMAC] Allow DMA to initialize when auto-negotiated to Gigabit
This commit is a bug-fix to the STMAC driver. When the STMAC is wired
directly to a Gigabit capable PHY, which was successfully able to
auto-negotiate to 1000Mbit on power-on, then the STMAC DMA will "hang"
during initialization. This problem is not present on 100Mbit networks, or
if the network is plugged in *after* initialization. Another finger-print
of this issue, is that, when the DMA initialization is "hanging" (as
the result of a "ping" or a "nfs" command), if the network connection
is temporarily broken, and re-connected, then it all works fine.
Subsequent network commands work fine, it is only the first network
command that hangs, if the PHY successfully auto-negotiated to Gigabit
speeds after power-on or reset.

Specifically, we need to initialize the PHY (and complete an
auto-negotiation) *before* we reset the MAC. This is needed in case the
PHY auto-negotiated to Gigabit, but the MAC is configured for the slower
100Mps (using MII).  This combination results in the lack of TX clock from
the PHY, (to the MAC) and can result in the GMAC not being resettable!
Specifically, the DMA reset code needs a good TX clk present.

This problem was noticed on a B2000 board, but could be present on all
boards with a Gigabit PHY on a Gigabit network, connected to the STMAC.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>